Add delete confirmation dialog

+++ pace-gui/src/main/java/edu/colorado/cires/pace/gui/MetadataForm.java
@@ -14,6 +14,7 @@ import java.awt.GridBagLayout;
 import java.util.UUID;
 import javax.swing.JCheckBox;
 import javax.swing.JLabel;
+import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JScrollPane;
 import javax.swing.JTextField;
@@ -91,7 +92,16 @@ public abstract class MetadataForm<O extends ObjectWithUniqueField> extends Form
   }
 
   @Override
-  protected void delete(CRUDRepository<O> repository) throws NotFoundException, DatastoreException, BadArgumentException {
-    repository.delete(UUID.fromString(uuid.getText()));
+  protected int delete(CRUDRepository<O> repository) throws NotFoundException, DatastoreException, BadArgumentException {
+    int result = JOptionPane.showConfirmDialog(this, "Are you sure you want to proceed? This action cannot be undone.", "Confirm Deletion", JOptionPane.YES_NO_OPTION);
+    if (result == JOptionPane.YES_OPTION) {
+      repository.delete(UUID.fromString(uuid.getText()));
+    } else if (result == JOptionPane.NO_OPTION) {
+      System.out.println("User chose NO");
+    } else {
+      System.out.println("User closed the dialog");
+    }
+
+    return result;
   }
 }
